requirements.txt
click==8.1.7
colorama==0.4.6
dlib 
face-recognition==1.3.0
face_recognition_models==0.3.0
filelock==3.15.4
fsspec==2024.6.1
intel-openmp==2021.4.0
Jinja2==3.1.4
MarkupSafe==2.1.5
mkl==2021.4.0
mkl-service==2.4.0
mpmath==1.3.0
networkx==3.2.1
numpy
opencv-python==4.10.0.84
pillow==10.4.0
sympy==1.12.1
tbb==2021.13.0
torch==2.3.1
typing_extensions==4.12.2
#conda install conda-forge::dlib
